[NFC][flang] Move `ReductionProcessor` to `Lower/Support`.
With #145837, the `ReductionProcessor` component is now used by both
OpenMP and `do concurrent`. Therefore, this PR moves it to a shared
location: `flang/Lower/Support`.
